pyproject.toml
[project]
name = "uni-esse3api"
version = "0.1.0"
description = ""
license="BSD-3-Clause"
readme = "README.md"
authors = [
    { name = "Francesco Filicetti", email = "francesco.filicetti@unical.it" }
]
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.10,<7.0",
    "djangorestframework>=3.16.1,<4.0",
    "drf-spectacular>=0.29.0,<1.0",
    "requests>=2.32.5,<3.0",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"
